instahyre/requirements.txt
drf-yasg
django
djangorestframework-simplejwt
psycopg2-binary
faker